fix review issues

+++ src/WorkOS.net/Entities/UpdateAuditLogsRetention.cs
@@ -9,9 +9,9 @@ namespace WorkOS
     public class UpdateAuditLogsRetention
     {
 
-        /// <summary>The number of days to retain audit logs. Valid values are 30 or 365.</summary>
+        /// <summary>The number of days Audit Log events will be retained. Valid values are `30` and `365`.</summary>
         [JsonProperty("retention_period_in_days")]
         [STJS.JsonPropertyName("retention_period_in_days")]
-        public double RetentionPeriodInDays { get; set; }
+        public long RetentionPeriodInDays { get; set; }
     }
 }
